ref(tests): Makes the test scaffolding an actual module
The current API tests made us export things from server that we didn't
need. This adds the test utils as an optional feature for people to import.

I also added some code for testing against a running server before I changed
gears to this implemention, so I kept it there for testing the client in a
follow up PR
