+++ Cargo.toml
@@ -27,10 +27,11 @@ exclude = [
 maintenance = { status = "actively-developed" }
 
 [features]
-default = ["server", "client", "caching"]
+default = ["server", "client", "caching", "test-tools"]
 server = ["warp"]
 client = ["reqwest", "mime_guess", "dirs"]
 caching = ["client"]
+test-tools = ["multipart"]
 cli = ["clap"]
 
 [dependencies]
@@ -61,9 +62,9 @@ log = "0.4.11"
 env_logger = "0.8"
 dirs = { version = "3.0", optional = true }
 mime_guess = { version = "2.0", optional = true }
+multipart = { version = "0.17", optional = true }
 
 [dev-dependencies]
-multipart = "0.17"
 mime = "0.3"
 
 [[bin]]
